% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/calculation.R
\name{swe_fixstar2_mag}
\alias{swe_fixstar2_mag}
\alias{swe_set_topo}
\alias{swe_calc}
\alias{swe_calc_ut}
\alias{swe_fixstar2}
\alias{swe_fixstar2_ut}
\title{Calculation magnitude of star}
\usage{
swe_fixstar2_mag(star)

swe_set_topo(geolon, geolat, altitude)

swe_calc(jd_et, ipl, iflag)

swe_calc_ut(jd_ut, ipl, iflag)

swe_fixstar2(star, jd_et, iflag)

swe_fixstar2_ut(star, jd_ut, iflag)
}
\arguments{
\item{geolon}{geographic longitude as double (deg)}

\item{geolat}{geographic latitude as double (deg)}

\item{altitude}{height as double (m)}
}
\value{
\code{swe_fixstar2_mag} returns a list with named entries \code{return} status flag as interger,
        \code{star} updated star name as string, \code{mag} magnitude of star as double, and \code{serr} for error message as string.

\code{swe_calc} returns a list with named entries \code{return} status flag as integer,
        \code{xx} updated planet name as string, and \code{serr} error message as string.

\code{swe_calc_ut} returns a list with named entries \code{return} status flag as integer,
        \code{xx} updated star name as string and \code{serr} error message as string.

\code{swe_fixstar2} returns a list with named entries \code{return} status flag as integer,
        \code{star} updated star name as string, \code{xx} star phenomena as numeric vector, and \code{serr} error message as string.

\code{swe_fixstar2_ut} returns a list with named entries \code{return} status flag as integer,
        \code{star} updated star name as string, \code{xx} star phenomena as numeric vector, and \code{serr} error message as string.
}
\description{
Calculation magnitude of star

Set the geographic location

Compute information of planet (ET)

Compute information of planet (UT)

Compute information of star (ET)

Compute information of star (UT)
}
