% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{day_of_week}
\alias{day_of_week}
\title{Compute day of week}
\usage{
day_of_week(jd)
}
\arguments{
\item{jd}{Julian date as double}
}
\value{
Day of week as integer
}
\description{
Compute day of week
}
