% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{deltat}
\alias{deltat}
\title{Determine the DeltaT at a certain date}
\usage{
deltat(tjd)
}
\arguments{
\item{tjd}{Julian day Number}
}
\value{
DeltaT (sec)
}
\description{
Determine the DeltaT at a certain date
}
