% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{swe_sol_eclipse_when_loc}
\alias{swe_sol_eclipse_when_loc}
\alias{swe_lun_eclipse_when_loc}
\alias{swe_lun_eclipse_how}
\alias{swe_lun_eclipse_when}
\alias{swe_rise_trans_true_hor}
\alias{swe_azalt}
\alias{swe_pheno_ut}
\alias{swe_pheno}
\alias{swe_heliacal_pheno_ut}
\alias{swe_topo_arcus_visionis}
\alias{swe_heliacal_angle}
\alias{swe_heliacal_ut}
\alias{swe_vis_limit_mag}
\title{Eclipses and planetary phenomena
Compute solar eclipse at location}
\usage{
swe_sol_eclipse_when_loc(tjd_start, ifl, geopos, backward)

swe_lun_eclipse_when_loc(tjd_start, ifl, geopos, backward)

swe_lun_eclipse_how(tjd_start, ifl, geopos)

swe_lun_eclipse_when(tjd_start, ifl, ifltype, backward)

swe_rise_trans_true_hor(tjd_ut, ipl, starname, epheflag, rsmi, geopos,
  atpress, attemp, horhgt)

swe_azalt(tjd_ut, calc_flag, geopos, atpress, attemp, xin)

swe_pheno_ut(tjd_ut, ipl, iflag)

swe_pheno(tjd_et, ipl, iflag)

swe_heliacal_pheno_ut(tjd_ut, dgeo, datm, dobs, objectname, event_type,
  helflag)

swe_topo_arcus_visionis(tjd_ut, dgeo, datm, dobs, helflag, mag, AziO, AltO,
  AziS, AziM, AltM)

swe_heliacal_angle(tjd_ut, dgeo, datm, dobs, helflag, mag, AziO, AziS,
  AziM, AltM)

swe_heliacal_ut(tjdstart, dgeo, datm, dobs, objectname, event_type,
  helflag)

swe_vis_limit_mag(tjd_ut, dgeo, datm, dobs, objectname, helflag)
}
\arguments{
\item{tjd_start}{Julian day, UT time}

\item{ifl}{Type of ephemeris (one of SEFLG_SWIEPH=2, SEFLG_JPLEPH=1, SEFLG_MOSEPH=4)}

\item{geopos}{The position vector (longitude, latitude, height)}

\item{backward}{TRUE for backwards search}

\item{ifltype}{Type of eclipse event (SE_ECL_TOTAL etc.  or 0, if any eclipse type)}

\item{tjd_ut}{Julian day, UT time}

\item{ipl}{body/planet number (-1 for no planet possible with \code{swe_rise_trans_true_hor})}

\item{starname}{star name ("" for no star)}

\item{epheflag}{Type of ephemeris (4=Moshier, 1=JPL, 2=SE)}

\item{rsmi}{integer specifying that rise, set, or one of the two meridian transits is wanted}

\item{atpress}{atmospheric pressure in mbar (hPa)}

\item{attemp}{atmospheric temperature in degrees Celsius}

\item{horhgt}{The apparent horizon at rise/set in degrees}

\item{calc_flag}{Type of reference system}

\item{xin}{position of body in either ecliptical or equatorial coordinates, depending on calc_flag}

\item{iflag}{a 32 bit integer containing bit flags that indicate what
kind of computation is wanted}

\item{tjd_et}{Julian day, Ephemeris time}

\item{dgeo}{Geographic position}

\item{datm}{Atmospheric conditions}

\item{dobs}{Observer description}

\item{objectname}{celectial object}

\item{event_type}{event type}

\item{helflag}{calcuation flag}

\item{mag}{The object's magnitude}

\item{AziO}{The object's azimut}

\item{AltO}{The object's altitude}

\item{AziS}{The sun's azimut}

\item{AziM}{The moon's azimut}

\item{AltM}{The moon's altitude}

\item{tjdstart}{Julian day, UT time}
}
\value{
\code{swe_rise_trans_true_hor} returns a list with named entries: \code{i} success of function
     \code{tret} for azi/alt info and \code{serr} for possible error code

\code{swe_azalt} returns a list with named entries:
     \code{xaz} for azi/alt info.

\code{swe_sol_eclipse_when_loc} returns a list with named entries: 
     \code{return} visibility code, \code{tret} for eclipse timing moments, 
     \code{attr} pheneomena durign eclipse and \code{serr} error string

\code{swe_lun_eclipse_when_loc} returns a list with named entries:
     \code{return} visibility code, \code{tret} for eclipse timing moments,
     \code{attr} pheneomena durign eclipse and \code{serr} error string

\code{swe_lun_eclipse_how} returns a list with named entries:
     \code{return} visibility code,
     \code{attr} pheneomena durign eclipse and \code{serr} error string

\code{swe_lun_eclipse_when} returns a list with named entries:
     \code{return} visibility code, \code{tret} for eclipse timing moments
     and \code{serr} error string

\code{swe_azalt} returns a list with named entries:
     \code{xaz} for azi/alt info.

\code{swe_pheno_ut} returns a list with named entries:
     \code{return} ???, \code{attr} for phenomenon information
     and \code{serr} error string

\code{swe_pheno} returns a list with named entries:
     \code{return} ???, \code{attr} for phenomenon information
     and \code{serr} error string

\code{swe_heliacal_pheno_ut} returns a list with named entries: \code{i} success of function
     \code{darr} for heliacal details and \code{serr} for possible error code

\code{swe_topo_arcus_visionis} returns a list with named entries: \code{i} success of function
     \code{darr} for heliacal details and \code{serr} for possible error code

\code{swe_heliacal_angle} returns a list with named entries: \code{i} success of function
     \code{darr} for heliacal angle and \code{serr} for possible error code

\code{swe_heliacal_ut} returns a list with named entries \code{return},
        \code{dret} results, and \code{serr} error message.

\code{swe_vis_limit_mag} returns a list with named entries: \code{i} success of function
     \code{dret} for magnitude info and \code{serr} for possible error code
}
\description{
Eclipses and planetary phenomena
Compute solar eclipse at location

Compute lunar eclipse at location

Computes the attributes of a lunar eclipse at a given time

Search a lunar eclipse on earth

Compute the rise and set location of the object

Compute horizon information: azimuth, altiiude

Provide phenomom information of celestial body (UT)

Provide phenomom information of celestial body (ET)

Compute heliacal event details

Compute heliacal event details

Compute heliacal event details

Compute the heliacale event of celestial object

Compute the limiting visibiliy magnitude
}
