% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bodies}
\alias{bodies}
\alias{flags}
\title{Enumerations used by the Swiss Ephemeris}
\format{A named list of integers.}
\source{
\url{http://www.astro.com/swisseph/swephprg.htm?lang=g#_Toc505244839}

\url{http://www.astro.com/swisseph/swephprg.htm?lang=g#_Toc505244840}
}
\usage{
bodies

flags
}
\description{
Integer constants representing all celestial bodies supported by the Swiss Ephemeris.

Integer flags used to change the calculations.
}
\keyword{datasets}
