% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_tid_acc}
\alias{get_tid_acc}
\title{Get the present configured tidal acceleration}
\usage{
get_tid_acc()
}
\value{
tidal acceleration as double (arcsec/century^2)
}
\description{
Get the present configured tidal acceleration
}
