% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{misc}
\alias{misc}
\alias{swe_set_ephe_path}
\alias{swe_close}
\alias{swe_version}
\alias{swe_deltat_ex}
\alias{swe_deltat}
\alias{swe_get_tid_acc}
\alias{swe_set_tid_acc}
\alias{swe_set_delta_t_userdef}
\alias{swe_day_of_week}
\alias{swe_julday}
\title{Miscellaneous functions}
\usage{
swe_set_ephe_path(path)

swe_close()

swe_version()

swe_deltat_ex(tjd, ephe_flag)

swe_deltat(tjd)

swe_get_tid_acc()

swe_set_tid_acc(t_acc)

swe_set_delta_t_userdef(delta_t)

swe_day_of_week(jd)

swe_julday(year, month, day, hour, gregflag)
}
\arguments{
\item{path}{the directory where the ephemeris files are stored (a string)}

\item{tjd}{Julian day Number}

\item{ephe_flag}{the epheemris flag (one of SEFLG_SWIEPH=2, SEFLG_JPLEPH=1, SEFLG_MOSEPH=4)}

\item{t_acc}{tidal acceleration as double (arcsec/century^2)}

\item{delta_t}{DeltaT value (sec)}

\item{jd}{Julian day as numeric vector}

\item{year}{Year}

\item{month}{Month}

\item{day}{Day}

\item{hour}{Hour}

\item{gregflag}{Gregorian (1) or Julian calandar (0)}
}
\value{
\code{swe_deltat} returns the DeltaT (sec)

\code{swe_day_of_week} retruns the day of week as integer vector (Monday = 0, ... Sunday = 6)

\code{swe_get_tid_acc} returns the tidal acceleration as double (arcsec/century^2)

\code{swe_version} returns the Swiss Ephemeris version number as string

\code{swe_deltat_ex} returns a list with named entries: \code{return} for return value
         and \code{serr} for error message.
}
\description{
Miscellaneous functions from the C API of the underlying
  \code{libswe}.
}
