% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEdata.R
\docType{data}
\name{SE}
\alias{SE}
\title{Constants used in swephR}
\format{
A data frame with 217 rows and 2 variables
}
\usage{
data(SE)
}
\description{
\itemize{
  \item name of variable
  \item value of the variable
}
}
\keyword{datasets}
