% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{swe_set_ephe_path}
\alias{swe_set_ephe_path}
\alias{swe_close}
\alias{swe_version}
\title{Set the directory for the sefstars.txt, swe_deltat.txt and jpl files}
\usage{
swe_set_ephe_path(path)

swe_close()

swe_version()
}
\description{
Set the directory for the sefstars.txt, swe_deltat.txt and jpl files

Close Swiss Ephemeris files

Get the Swiss Ephemeris version number
}
