% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Section14}
\alias{Section14}
\alias{swe_house_pos}
\title{Section 14: House position calculations}
\usage{
swe_house_pos(armc, geolat, eps, hsys, xpin)
}
\arguments{
\item{armc}{right ascension of the MC as double (deg)}

\item{geolat}{geographic latitude as double (deg)}

\item{eps}{ecliptic obliquity as double (deg)}

\item{hsys}{house method, one-letter case sensitive as char}

\item{xpin}{longitude and latitude of the given boby as numeric vector (deg)}
}
\value{
\code{swe_house_pos} returns a list with named entries: \code{return} how far from body's cusp as double,
     and \code{serr} error message as string.
}
\description{
Calculate house position of a given body.
}
\details{
\describe{
\item{swe_house_pos()}{Calculate house position of given body.}
}
}
\examples{
swe_house_pos(12, 53, 23, 'B', c(0,0))
}
\seealso{
Section 14 in \url{http://www.astro.com/swisseph/swephprg.htm}
}
