% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Section15}
\alias{Section15}
\alias{swe_sidtime}
\title{Section 15: Sidereal time}
\usage{
swe_sidtime(jd_ut)
}
\arguments{
\item{jd_ut}{UT Julian day number as double (day)}
}
\value{
\code{swe_sidtime} returns the sidereal time as double (deg)
}
\description{
Calculate the sidereal time (in degrees).
}
\details{
\describe{
\item{swe_sidtime()}{Determine the sidereal time.}
}
}
\examples{
swe_sidtime(2451545)
}
\seealso{
Section 15 in \url{http://www.astro.com/swisseph/swephprg.htm}. Remember that array indices start in R at 1, while in C they start at 0!
}
