% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Section2.R
\name{swe_calc}
\alias{swe_calc}
\alias{swe_calc_ut}
\title{Compute information of planet (ET)}
\usage{
swe_calc(jd_et, ipl, iflag)

swe_calc_ut(jd_ut, ipl, iflag)
}
\description{
Compute information of planet (ET)

Compute information of planet (UT)
}
