% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{swe_get_planet_name}
\alias{swe_get_planet_name}
\title{Section 3: Find a planetary or asteroid name}
\usage{
swe_get_planet_name(ipl)
}
\arguments{
\item{ipl}{body/planet as interger (SE_SUN=0, SE_Moon=1,  ... SE_PLUTO=9)}
}
\value{
\code{swe_get_planet_name} returns objectname as string
}
\description{
Find a planetary or asteroid name.
}
\details{
\describe{
  \item{swe_get_planet_name()}{Convert object number into ubject name.}
}
}
\examples{
swe_get_planet_name(SE_MOON)
}
\seealso{
\url{http://www.astro.com/swisseph/swephprg.htm#_Toc505244843}
}
