% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/Section4.R
\name{swe_fixstar2_mag}
\alias{swe_fixstar2_mag}
\alias{swe_fixstar2}
\alias{swe_fixstar2_ut}
\title{Calculation magnitude of star}
\usage{
swe_fixstar2_mag(starname)

swe_fixstar2(star, jd_et, iflag)

swe_fixstar2_ut(star, jd_ut, iflag)
}
\value{
\code{swe_fixstar2_mag} returns a list with named entries \code{return} status flag as interger,
        \code{starname} updated star name as string, \code{mag} magnitude of star as double, and \code{serr} for error message as string.

\code{swe_fixstar2} returns a list with named entries \code{return} status flag as integer,
        \code{star} updated star name as string, \code{xx} star phenomena as numeric vector, and \code{serr} error message as string.

\code{swe_fixstar2_ut} returns a list with named entries \code{return} status flag as integer,
        \code{star} updated star name as string, \code{xx} star phenomena as numeric vector, and \code{serr} error message as string.
}
\description{
Calculation magnitude of star

Compute information of star (ET)

Compute information of star (UT)
}
