% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{swe_sol_eclipse_when_loc}
\alias{swe_sol_eclipse_when_loc}
\alias{swe_lun_eclipse_when_loc}
\alias{swe_lun_eclipse_how}
\alias{swe_lun_eclipse_when}
\alias{swe_rise_trans_true_hor}
\alias{swe_azalt}
\alias{swe_azalt_rev}
\alias{swe_pheno_ut}
\alias{swe_pheno}
\alias{swe_heliacal_pheno_ut}
\alias{swe_topo_arcus_visionis}
\alias{swe_heliacal_angle}
\alias{swe_heliacal_ut}
\alias{swe_vis_limit_mag}
\alias{swe_refrac_extended}
\title{Section 6: Eclipses, Risings, Settings, Meridian Transits, Planetary Phenomena}
\usage{
swe_sol_eclipse_when_loc(jd_start, ephe_flag, geopos, backward)

swe_lun_eclipse_when_loc(jd_start, ephe_flag, geopos, backward)

swe_lun_eclipse_how(jd_start, ephe_flag, geopos)

swe_lun_eclipse_when(jd_start, ephe_flag, ifltype, backward)

swe_rise_trans_true_hor(jd_ut, ipl, star, ephe_flag, rsmi, geopos, atpress,
  attemp, horhgt)

swe_azalt(jd_ut, calc_flag, geopos, atpress, attemp, xin)

swe_azalt_rev(jd_ut, calc_flag, geopos, xin)

swe_pheno_ut(jd_ut, ipl, ephe_flag)

swe_pheno(jd_et, ipl, ephe_flag)

swe_heliacal_pheno_ut(jd_ut, dgeo, datm, dobs, objectname, event_type,
  helflag)

swe_topo_arcus_visionis(jd_ut, dgeo, datm, dobs, helflag, mag, AziO, AltO,
  AziS, AziM, AltM)

swe_heliacal_angle(jd_ut, dgeo, datm, dobs, helflag, mag, AziO, AziS, AziM,
  AltM)

swe_heliacal_ut(jd_utstart, dgeo, datm, dobs, objectname, event_type,
  helflag)

swe_vis_limit_mag(jd_ut, dgeo, datm, dobs, objectname, helflag)

swe_refrac_extended(InAlt, geoheight, atpress, attemp, lapse_rate,
  calc_flag)
}
\arguments{
\item{jd_start}{Julian day number (UT)}

\item{ephe_flag}{ephemeris flag as integer (SEFLG_JPLEPH=1, SEFLG_SWIEPH=2 or SEFLG_MOSEPH=4)}

\item{geopos}{The position numeric vector (longitude, latitude, height)}

\item{backward}{backwards search as boolean (TRUE)}

\item{ifltype}{eclipse tper as interger (e.g.: SE_ECL_CENTRAL=1,SE_ECL_NONCENTRAL=2,SE_ECL_TOTAL=4,SE_ECL_ANNULAR=8,SE_ECL_PARTIAL=16,SE_ECL_ANNULAR_TOTAL=32)}

\item{jd_ut}{UT Julian day number (day)}

\item{ipl}{body/planet as interger (SE_SUN=0, SE_Moon=1,  ... SE_PLUTO=9)}

\item{star}{star name as string ("" for no star)}

\item{rsmi}{event flag as integer (e.g.: SE_CALC_RISE=1, SE_CALC_SET=2,SE_CALC_MTRANSIT=4,SE_CALC_ITRANSIT=8)}

\item{atpress}{atmospheric pressure as double (hPa)}

\item{attemp}{atmospheric temperature as double (Celsius)}

\item{horhgt}{horizon apparent altitude as double (deg)}

\item{calc_flag}{flag as interger: reference system (e.g.: SEFLG_EQUATORIAL    2048 or ecliptic) or refraction direction (SE_TRUE_TO_APP=0, SE_APP_TO_TRUE=1)}

\item{xin}{position of body as numeric vector (either ecliptical or equatorial coordinates, depending on calc_flag)}

\item{jd_et}{ET Julian day number as double (day)}

\item{dgeo}{Geographic position as numeric vector}

\item{datm}{Atmospheric conditions as numeric vector}

\item{dobs}{Observer description as numeric vector}

\item{objectname}{celectial object as string}

\item{event_type}{event type as integer}

\item{helflag}{calcuation flag as integer}

\item{mag}{object's magnitude as double (-)}

\item{AziO}{object's azimuth as double (deg)}

\item{AltO}{object's altitude as double (deg)}

\item{AziS}{Sun's azimuth as double (deg)}

\item{AziM}{Moon's azimut as double (deg)}

\item{AltM}{Moon's altitude as double (deg)}

\item{jd_utstart}{UT Julian day number as double (day)}

\item{InAlt}{object's apparent/topocentric altitude as double (depending on calc_flag) (deg)}

\item{geoheight}{observer's height as double (m)}

\item{lapse_rate}{lapse rate as double (K/m)}
}
\value{
\code{swe_rise_trans_true_hor} returns a list with named entries: \code{return} status flag as interger,
     \code{tret} for azi/alt info as numeric vector and \code{serr} error message as string

\code{swe_azalt} returns a list with named entries:
     \code{xaz} for azi/alt info as numeric vector.

\code{swe_sol_eclipse_when_loc} returns a list with named entries: 
     \code{return} status flag as interger, \code{tret} for eclipse timing moments as numeric vector, 
     \code{attr} pheneomena during eclipse as numeric vector and \code{serr} error message as string

\code{swe_lun_eclipse_when_loc} returns a list with named entries:
     \code{return} status flag as integer, \code{tret} for eclipse timing moments,
     \code{attr} pheneomena during eclipse and \code{serr} error warning as string

\code{swe_lun_eclipse_how} returns a list with named entries:
     \code{return} status flag as integer,
     \code{attr} pheneomena during eclipse as numeric vector and \code{serr} error message as string

\code{swe_lun_eclipse_when} returns a list with named entries:
     \code{return} status flag as interger, \code{tret} for eclipse timing moments as numeric vector
     and \code{serr} error warning as string

\code{swe_azalt} returns a list with named entries:
     \code{xaz} for azi/alt info as numeric vector.

\code{swe_azalt_rev} returns a list with named entries:
     \code{xaz} for celestial info a snumeric vector.

\code{swe_pheno_ut} returns a list with named entries:
     \code{return} status fag as integer, \code{attr} for phenomenon information as numeric vector
     and \code{serr} error warning as string

\code{swe_pheno} returns a list with named entries:
     \code{return} status fag as integer, \code{attr} for phenomenon information as numeric vector
     and \code{serr} error message as string

\code{swe_heliacal_pheno_ut} returns a list with named entries: \code{return} status flag as integer
     \code{darr} for heliacal details as numeric vector and \code{serr} error message as string

\code{swe_topo_arcus_visionis} returns a list with named entries: \code{return} status flag as integer,
     \code{darr} heliacal details as numeric vector and \code{serr} error message as string

\code{swe_heliacal_angle} returns a list with named entries: \code{return} status flag as integer,
     \code{dret} heliacal angle as numeric vector and \code{serr} error message as string

\code{swe_heliacal_ut} returns a list with named entries \code{return} status flag as integer,
        \code{dret} heliacal results as numeric vector, and \code{serr} error message as string.

\code{swe_vis_limit_mag} returns a list with named entries: \code{return} status flag as interger,
     \code{dret} limiting magnitude as double and \code{serr} error message as string

\code{swe_refrac_extended} returns a list with named entries: \code{return} status flag as interger,
     \code{dret} refraction results as nemeric vector (TopoAlt, AppAlt, refraction)
}
\description{
swe_lun_eclipse_when_loc: This function determines the moment a lunar eclipse event happens.
}
