% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{swe_set_topo}
\alias{swe_set_topo}
\title{Section 9: The function for topocentric planet positions}
\usage{
swe_set_topo(geolon, geolat, altitude)
}
\arguments{
\item{geolon}{geographic longitude as double (deg)}

\item{geolat}{geographic latitude as double (deg)}

\item{altitude}{height as double (m)}
}
\description{
Section 9: The function for topocentric planet positions
}
