% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SectionConstants.R
\name{SEConstants}
\alias{SEConstants}
\title{Initialising Constants: Initialising constants used for the SE function calls.}
\usage{
SEConstants()
}
\value{
\code{SEConstants} returns a list with named entries: the constants in SE
}
\description{
The constants defined in SE are in R defined as a list.
If the function return is stored in the variable SE than for instance SE_MOON becomes SE$MOON
and SEFLG_JPLEPH becomes SE$FLG_JPLEPH.
}
\details{
\describe{
  \item{SEConstants()}{Initialise SE constants in R.}
 }
}
\examples{
SEConstants()
}
\seealso{
\url{http://www.astro.com/swisseph/swephprg.htm}
}
