% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_delta_t_userdef}
\alias{set_delta_t_userdef}
\title{Set one's own DeltaT}
\usage{
set_delta_t_userdef(t_acc)
}
\arguments{
\item{t_acc}{DeltaT value (sec)
return void}
}
\description{
Set one's own DeltaT
}
