% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_ephe_path}
\alias{set_ephe_path}
\title{Set the directory for the sefstar.txt, swe_deltat.txt and jpl files}
\usage{
set_ephe_path(file)
}
\arguments{
\item{file}{the directory plus file (a string)}
}
\description{
Set the directory for the sefstar.txt, swe_deltat.txt and jpl files
}
