% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_tid_acc}
\alias{set_tid_acc}
\title{Set the tidal acceleration}
\usage{
set_tid_acc(t_acc)
}
\arguments{
\item{t_acc}{tidal acceleration as double (arcsec/century^2)}
}
\description{
Set the tidal acceleration
}
