% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_topo}
\alias{set_topo}
\title{Set the topocentric location (lon, lat, height)}
\usage{
set_topo(geolon, geolat, altitude)
}
\arguments{
\item{geolon}{Topocentric Longitude (deg)}

\item{geolat}{Topocentric Latitude (deg)}

\item{altitude}{the height (m)}
}
\description{
Set the topocentric location (lon, lat, height)
}
