% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swephR-package.R
\docType{package}
\name{swephR-package}
\alias{swephR}
\alias{swephR-package}
\title{swephR: High Precision Swiss Ephemeris}
\description{
The Swiss Ephemeris is a high precision ephemeris developed
    by Astrodienst, based upon the DE431 ephemerides from NASA's JPL. This
    package uses the semianalytic theory by Steve Moshier. For faster and more
    accurate calculations, the compressed Swiss Ephemeris data is available in
    the 'swephRdata' package. To access this data package, run
    'install.packages("swephRdata", repos = "https://rstub.github.io/drat/",
    type = "source")'. The size of the 'swephRdata' package is approximately
    115 MB.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstub/swephR/}
  \item \url{http://www.astro.com/swisseph/swephinfo_e.htm}
  \item Report bugs at \url{https://github.com/rstub/swephR/issues/}
}

}
