% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwissEphemeris-package.R
\docType{package}
\name{SwissEphemeris-package}
\alias{SwissEphemeris}
\alias{SwissEphemeris-package}
\title{R interface to the Swiss Ephemeris}
\description{
The SwissEphemeris package provides an R interface for the
\href{http://www.astro.com/swisseph/swephinfo_e.htm}{\code{Swiss Ephemeris}}.
}
