% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorisedSEfunctions.R
\name{vec_calc_ut}
\alias{vec_calc_ut}
\alias{vec_calc}
\alias{vec_get_planet_name}
\alias{vec_sol_eclipse_when_loc}
\alias{vec_fixstar2_ut}
\alias{vec_fixstar2}
\alias{vec_fixstar2_mag}
\alias{vec_lun_eclipse_when_loc}
\alias{vec_lun_eclipse_how}
\alias{vec_lun_eclipse_when}
\alias{vec_rise_trans_true_hor}
\alias{vec_pheno_ut}
\alias{vec_pheno}
\alias{vec_azalt}
\alias{vec_azalt_rev}
\alias{vec_refrac_extended}
\alias{vec_heliacal_ut}
\alias{vec_vis_limit_mag}
\alias{vec_heliacal_pheno_ut}
\alias{vec_topo_arcus_visionis}
\alias{vec_heliacal_angle}
\alias{vec_julday}
\alias{vec_date_conversion}
\alias{vec_revjul}
\alias{vec_deltat_ex}
\alias{vec_deltat}
\alias{vec_day_of_week}
\title{Vectorised swephR functions}
\usage{
vec_calc_ut(jd_ut, ipl, iflag = 4)

vec_calc(jd_et, ipl, iflag = 4)

vec_get_planet_name(ipl)

vec_sol_eclipse_when_loc(jd_start, ephe_flag = 4, long, lat,
  height = 0, backward = FALSE)

vec_fixstar2_ut(starname, jd_ut, iflag = 4)

vec_fixstar2(starname, jd_et, iflag = 4)

vec_fixstar2_mag(starname)

vec_lun_eclipse_when_loc(jd_start, ephe_flag = 4, long, lat,
  height = 0, backward = FALSE)

vec_lun_eclipse_how(jd_ut, ephe_flag = 4, long, lat, height = 0)

vec_lun_eclipse_when(jd_start, ephe_flag = 4, ifltype,
  backward = FALSE)

vec_rise_trans_true_hor(jd_ut, ipl, starname = "", ephe_flag = 4, rsmi,
  long, lat, height = 0, atpress = 1013.25, attemp = 15,
  horhgt = 0)

vec_pheno_ut(jd_ut, ipl, iflag = 4)

vec_pheno(jd_et, ipl, iflag = 4)

vec_azalt(jd_ut, coord_flag, long, lat, height = 0, atpress = 1013.25,
  attemp = 15, xin1, xin2)

vec_azalt_rev(jd_ut, coord_flag, long, lat, height = 0, xin1, xin2)

vec_refrac_extended(InAlt, height = 0, atpress = 1013.25,
  attemp = 15, lapse_rate = 0.0065, calc_flag)

vec_heliacal_ut(jd_utstart, long, lat, height = 0, atpress = 1013.25,
  attemp = 15, athum = 0, atvis = 0.2, obsage = 36,
  obssnellen = 1.4, obsbin = 1, obsmag = 1, obsaper = 7,
  obstrans = 0.8, objectname, event_type, helflag = 768)

vec_vis_limit_mag(jd_ut, long, lat, height = 0, atpress = 1013.25,
  attemp = 15, athum = 0, atvis = 0.2, obsage = 36,
  obssnellen = 1.4, obsbin = 1, obsmag = 1, obsaper = 7,
  obstrans = 0.8, objectname, helflag = 768)

vec_heliacal_pheno_ut(jd_ut, long, lat, height = 0, atpress = 1013.25,
  attemp = 15, athum = 0, atvis = 0.2, obsage = 36,
  obssnellen = 1.4, obsbin = 1, obsmag = 1, obsaper = 7,
  obstrans = 0.8, objectname, event_type, helflag = 768)

vec_topo_arcus_visionis(jd_ut, long, lat, height = 0,
  atpress = 1013.25, attemp = 15, athum = 0, atvis = 0.2,
  obsage = 36, obssnellen = 1.4, obsbin = 1, obsmag = 1,
  obsaper = 7, obstrans = 0.8, helflag = 768, mag, AziO, AltO, AziS,
  AziM, AltM)

vec_heliacal_angle(jd_ut, long, lat, height = 0, atpress = 1013.25,
  attemp = 15, athum = 0, atvis = 0.2, obsage = 36,
  obssnellen = 1.4, obsbin = 1, obsmag = 1, obsaper = 7,
  obstrans = 0.8, helflag = 768, mag, AziO, AziS, AziM, AltM)

vec_julday(year, month, day, hour = 12, gregflag = 1)

vec_date_conversion(year, month, day, hour = 12, cal = "g")

vec_revjul(jd, gregflag = 1)

vec_deltat_ex(jd_ut, ephe_flag = 4)

vec_deltat(jd_ut)

vec_day_of_week(jd)
}
\description{
This are the vectorised functions of the 
ones that are namend swe_xxx. Each input parameter can be a vector.
Make sure that the vector length of each input parameter matches 
the rules around data-frames: so any vector length must be an integer 
amount of the largest vector length.
}
