% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/calculation.R
\name{swe_fixstar2_mag}
\alias{swe_fixstar2_mag}
\alias{swe_set_topo}
\alias{swe_calc}
\alias{swe_calc_ut}
\alias{swe_fixstar2}
\alias{swe_fixstar2_ut}
\title{Calculation magnitude of star}
\usage{
swe_fixstar2_mag(star)

swe_set_topo(geolon, geolat, altitude)

swe_calc(tjd_et, ipl, iflag)

swe_calc_ut(tjd_ut, ipl, iflag)

swe_fixstar2(star, tjd_et, iflag)

swe_fixstar2_ut(star, tjd_ut, iflag)
}
\arguments{
\item{geolon}{geographic longitude as double (deg)}

\item{geolat}{geographic latitude as double (deg)}

\item{altitude}{height as double (m)}
}
\value{
\code{swe_fixstar2_mag} returns a list with named entries \code{return} status flag as interger,
        \code{star} updated star name as string, \code{mag} magnitude of star as double, and \code{serr} for error message as string.

\code{swe_calc} returns a list with named entries \code{rc},
        \code{xx} updated planet name, and \code{serr} error message.

\code{swe_calc_ut} returns a list with named entries \code{rc},
        \code{xx} updated star name, and \code{serr} error message.

\code{swe_fixstar2} returns a list with named entries \code{return},
        \code{star} updated star name, \code{xx}, and \code{serr} error message.

\code{swe_fixstar2_ut} returns a list with named entries \code{return},
        \code{star} updated star name, \code{xx}, and \code{serr} error message.
}
\description{
Calculation magnitude of star

Set the geographic location

Compute information of planet (ET)

Compute information of planet (UT)

Compute information of star

Compute information of star
}
