% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R, R/RcppExports.R
\name{calculation}
\alias{calculation}
\alias{swe_calc}
\alias{swe_calc_ut}
\alias{swe_fixstar2}
\alias{swe_fixstar2_ut}
\alias{swe_fixstar2_mag}
\alias{swe_set_topo}
\title{Calculation of planets and stars}
\usage{
swe_calc(tjd_et, ipl, iflag)

swe_calc_ut(tjd_ut, ipl, iflag)

swe_fixstar2(star, tjd_et, iflag)

swe_fixstar2_ut(star, tjd_ut, iflag)

swe_fixstar2_mag(star)

swe_set_topo(geolon, geolat, altitude)
}
\arguments{
\item{tjd_et}{Julian day, Ephemeris time}

\item{ipl}{body/planet number (-1 for no planet possible with \code{swe_rise_trans_true_hor})}

\item{iflag}{a 32 bit integer containing bit flags that indicate what
kind of computation is wanted}

\item{tjd_ut}{Julian day, UT time}

\item{star}{star name}

\item{geolon}{Topocentric Longitude (deg)}

\item{geolat}{Topocentric Latitude (deg)}

\item{altitude}{the height (m)}
}
\value{
\code{swe_calc} returns a list with named entries \code{rc},
        \code{xx} updated planet name, and \code{serr} error message.

\code{swe_calc_ut} returns a list with named entries \code{rc},
        \code{xx} updated star name, and \code{serr} error message.

\code{swe_fixstar2} returns a list with named entries \code{return},
        \code{star} updated star name, \code{xx}, and \code{serr} error message.

\code{swe_fixstar2_ut} returns a list with named entries \code{return},
        \code{star} updated star name, \code{xx}, and \code{serr} error message.

\code{swe_fixstar2_mag} returns a list with named entries \code{return},
        \code{star} updated star name, \code{mag} magnitude of star, and \code{serr} for error message.
}
\description{
Calculation of planets and stars

Compute information of planet (ET)

Compute information of planet (UT)

Compute information of star

Compute information of star
}
