% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expert-interface}
\alias{expert-interface}
\alias{swe_day_of_week}
\alias{swe_get_tid_acc}
\alias{swe_version}
\alias{swe_set_tid_acc}
\alias{swe_deltat}
\alias{swe_set_ephe_path}
\alias{swe_set_topo}
\alias{swe_set_delta_t_userdef}
\alias{swe_calc}
\title{Expert interface}
\usage{
swe_day_of_week(jd)

swe_get_tid_acc()

swe_version()

swe_set_tid_acc(t_acc)

swe_deltat(tjd)

swe_set_ephe_path(file)

swe_set_topo(geolon, geolat, altitude)

swe_set_delta_t_userdef(delta_t)

swe_calc(tjd_et, ipl, iflag = 0L)
}
\arguments{
\item{jd}{Julian date as numeric vector}

\item{t_acc}{tidal acceleration as double (arcsec/century^2)}

\item{tjd}{Julian day Number}

\item{file}{the directory plus file (a string)}

\item{geolon}{Topocentric Longitude (deg)}

\item{geolat}{Topocentric Latitude (deg)}

\item{altitude}{the height (m)}

\item{delta_t}{DeltaT value (sec)}

\item{tjd_et}{Julian day, Ephemeris time}

\item{ipl}{body number}

\item{iflag}{a 32 bit integer containing bit flags that indicate what
kind of computation is wanted}
}
\value{
\code{swe_day_of_week} retruns the day of week as integer vector

tidal acceleration as double (arcsec/century^2)

Swiss Ephemeris version number as string

DeltaT (sec)

\code{swe_calc} returns a list with named entries \code{rc},
        \code{xx}, and \code{serr} for return code, calculated values
        and error message.
}
\description{
For experts that are familiar with the C API of the underlying \code{libswe},
an expert interface that mimicks that interface as closely as possbile is
provided.
}
