% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expert-interface}
\alias{expert-interface}
\alias{swe_day_of_week}
\alias{swe_get_tid_acc}
\alias{swe_version}
\alias{swe_set_tid_acc}
\alias{swe_deltat}
\alias{swe_set_ephe_path}
\alias{swe_julday}
\alias{swe_set_delta_t_userdef}
\alias{swe_deltat_ex}
\alias{swe_close}
\title{Expert interface}
\usage{
swe_day_of_week(jd)

swe_get_tid_acc()

swe_version()

swe_set_tid_acc(t_acc)

swe_deltat(tjd)

swe_set_ephe_path(path)

swe_julday(year, month, day, hour, gregflag)

swe_set_delta_t_userdef(delta_t)

swe_deltat_ex(tjd, ephe_flag)

swe_close()
}
\arguments{
\item{jd}{Julian date as numeric vector}

\item{t_acc}{tidal acceleration as double (arcsec/century^2)}

\item{tjd}{Julian day Number}

\item{path}{the directory where the ephemeris files are stored (a string)}

\item{year}{Year}

\item{month}{Month}

\item{day}{Day}

\item{hour}{Hour}

\item{gregflag}{Greogiran (1) or juian calandar (0)}

\item{delta_t}{DeltaT value (sec)}

\item{ephe_flag}{the epheemris flag (one of SEFLG_SWIEPH=2, SEFLG_JPLEPH=1, SEFLG_MOSEPH=4)}
}
\value{
\code{swe_deltat} returns the DeltaT (sec)

\code{swe_day_of_week} retruns the day of week as integer vector

\code{get_tid_acc} returns the tidal acceleration as double (arcsec/century^2)

\code{swe_version} returns the Swiss Ephemeris version number as string

\code{swe_deltat_ex} returns a list with named entries: \code{return} for return value
         and \code{serr} for error message.
}
\description{
For experts that are familiar with the C API of the underlying
  \code{libswe}, an expert interface that mimicks that interface as closely
  as possbile is provided.
}
