% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Section0.R
\name{SEtest}
\alias{SEtest}
\title{Section 0: The programming steps to get planet position}
\usage{
SEtest()
}
\description{
A miniature sample program to calcuate the planet's positions.
}
\details{
\describe{
  \item{SEtest()}{Calcuate the planets' positions.} 
 }
}
\examples{
\dontrun{
SEtest()
}
}
\seealso{
\url{http://www.astro.com/swisseph/swephprg.htm#_Toc505244830}
}
