% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{swe_set_ephe_path}
\alias{swe_set_ephe_path}
\alias{swe_close}
\alias{swe_version}
\title{Section 1: The Ephemeris file related functions}
\usage{
swe_set_ephe_path(path)

swe_close()

swe_version()
}
\arguments{
\item{path}{directory for the sefstars.txt, swe_deltat.txt and jpl files}
}
\description{
Several initialisation and generic functions
}
