% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{swe_version}
\alias{swe_version}
\alias{swe_set_ephe_path}
\alias{swe_close}
\title{Get the Swiss Ephemeris version number}
\usage{
swe_version()

swe_set_ephe_path(path)

swe_close()
}
\description{
Get the Swiss Ephemeris version number

Set the directory for the sefstars.txt, swe_deltat.txt and jpl files

Close Swiss Ephemeris files
}
