% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Section16}
\alias{Section16}
\alias{swe_day_of_week}
\title{Section 16.7: Other functions that may be useful}
\usage{
swe_day_of_week(jd)
}
\arguments{
\item{jd}{Julian day number as numeric vector (day)}
}
\value{
\code{swe_day_of_week} retruns the day of week as integer vector (0 Monday .. 6 Sunday)
}
\description{
Useful functions
}
\details{
\describe{
  \item{we_day_of_week()}{Determine day of week from Julian day number.}
}
}
\examples{
swe_day_of_week(1234.567)
}
\seealso{
\url{http://www.astro.com/swisseph/swephprg.htm#_Toc505244906}
}
