% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Section2.R
\name{swe_calc}
\alias{swe_calc}
\alias{swe_calc_ut}
\title{Section 2: Computing positions}
\usage{
swe_calc(jd_et, ipl, iflag)

swe_calc_ut(jd_ut, ipl, iflag)
}
\arguments{
\item{jd_et}{ET Julian day number as double (day)}

\item{ipl}{body/planet as interger (SE_SUN=0, SE_Moon=1,  ... SE_PLUTO=9)}

\item{iflag}{flag as interger, many options possible (section 2.3)}

\item{jd_ut}{UT Julian day number as double (day)}
}
\value{
\code{swe_calc} returns a list with named entries \code{return} status flag as integer,
        \code{xx} updated star name as string and \code{serr} error message as string.

\code{swe_calc_ut} returns a list with named entries \code{return} status flag as integer,
       \code{xx} updated planet name as string, and \code{serr} error message as string.
}
\description{
Computing positions of planets, asteroids, lunar nodes and apogees using Swiss Ephemeris.
}
\details{
swe_calc()

It compute positions using ET.

swe_calc_ut()

It compute positions using UT.
}
