% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Section3}
\alias{Section3}
\alias{swe_get_planet_name}
\title{Section 3: Find a planetary or asteroid name}
\usage{
swe_get_planet_name(ipl)
}
\arguments{
\item{ipl}{Body/planet as integer (SE$SUN=0, SE$Moon=1,  ... SE$PLUTO=9)}
}
\value{
\code{swe_get_planet_name} returns the object's name as string
}
\description{
Find a planetary or asteroid name.
}
\details{
\describe{
  \item{swe_get_planet_name()}{Convert object number into object name.}
}
}
\examples{
data(SE)
swe_get_planet_name(SE$MOON)
}
\seealso{
Section 3 in \url{http://www.astro.com/swisseph/swephprg.htm}. Remember that array indices start in R at 1, while in C they start at 0!
}
