% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{swe_julday}
\alias{swe_julday}
\title{Section 7: Date and time conversion functions}
\usage{
swe_julday(year, month, day, hour, gregflag)
}
\arguments{
\item{year}{Year as interger}

\item{month}{Month as interger}

\item{day}{Day as interger}

\item{hour}{Hour as double}

\item{gregflag}{calendar type (SE_JUL_CAL=0 or SE_GREG_CAL=1)}
}
\description{
Functions related to calendar and time conversions
}
\details{
swe_julday()

Convert calendar dates to the astronomical time scale which measures time in Julian day number
}
