% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{swe_julday}
\alias{swe_julday}
\title{Section 7: Date and time conversion functions}
\usage{
swe_julday(year, month, day, hour, gregflag)
}
\arguments{
\item{year}{Year as integer}

\item{month}{Month as integer}

\item{day}{Day as integer}

\item{hour}{Hour as double}

\item{gregflag}{Calendar type (SE$JUL_CAL=0 or SE$GREG_CAL=1)}
}
\description{
Functions related to calendar and time conversions.
}
\details{
\describe{
  \item{swe_julday()}{Convert calendar dates to the astronomical time scale which measures time in Julian day number.}
}
}
\examples{
SE<-SEConstants()
swe_julday(2000,1,1,12,SE$GREG_CAL)
}
\seealso{
\url{http://www.astro.com/swisseph/swephprg.htm#_Toc505244873}
}
