% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{swe_get_tid_acc}
\alias{swe_get_tid_acc}
\alias{swe_set_tid_acc}
\alias{swe_deltat}
\alias{swe_set_delta_t_userdef}
\alias{swe_deltat_ex}
\title{Get the present configured tidal acceleration}
\usage{
swe_get_tid_acc()

swe_set_tid_acc(t_acc)

swe_deltat(jd_ut)

swe_set_delta_t_userdef(delta_t)

swe_deltat_ex(jd_ut, ephe_flag)
}
\arguments{
\item{ephe_flag}{ephemeris flag as integer (SEFLG_JPLEPH=1, SEFLG_SWIEPH=2 or SEFLG_MOSEPH=4) (section 2.3.2)}
}
\value{
\code{swe_deltat_ex} returns a list with named entries: \code{deltat} for DeltaT as double (day)
         and \code{serr} for error message as string.
}
\description{
Get the present configured tidal acceleration

Set the tidal acceleration

Determine the DeltaT at a certain date

Set one's own DeltaT

Determine DeltaT
}
