% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Section9}
\alias{Section9}
\alias{swe_set_topo}
\title{Section 9: The function for calculating topocentric planet position}
\usage{
swe_set_topo(longitude, lat, height)
}
\arguments{
\item{longitude}{Geographic longitude as double (deg)}

\item{lat}{Geographic latitude as double (deg)}

\item{height}{Height as double (m)}
}
\description{
Function for topocentric planet positions
}
\details{
\describe{
\item{swe_set_topo()}{Set the topocentric location of the observer.}
}
}
\examples{
swe_set_topo(0,50,10)
}
\seealso{
Section 9 in \url{http://www.astro.com/swisseph/swephprg.htm}. Remember that array indices start in R at 1, while in C they start at 0!
}
