% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swephR-package.R
\docType{package}
\name{swephR-package}
\alias{swephR}
\alias{swephR-package}
\title{R interface to the Swiss Ephemeris}
\description{
The swephR package provides an R interface for the
\href{http://www.astro.com/swisseph/swephinfo_e.htm}{\code{Swiss Ephemeris}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstub/swephR/}
}

}
